refactor: split tags tree component into files
